Crate wasmtime_wast

Implementation of the WAST text format for wasmtime.

Structs

SpectestConfig

Configuration of how spectest primitives work.

WastContext

The wast test script language allows modules to be defined and actions to be performed on them.

Constants

VERSION

Version number of this crate.

Functions

link_spectest

Return an instance implementing the “spectest” interface used in the spec testsuite.